--- src/components/Alert/Alert.module.scss
@@ -1,63 +1,0 @@
-@import "../../../src/theme/old-tokens/pallete";
-@import "../../../src/theme/old-tokens/functions";
-@import "../../../src/theme/old-tokens/typography";
-
-$alerts: (
-  "default": (
-    "backgroundColor": scale-color($green, $lightness: 80%),
-    "titleColor": #555,
-    "buttonColor": $green,
-  ),
-  "info": (
-    "backgroundColor": scale-color($info, $lightness: 60%),
-    "titleColor": #555,
-    "buttonColor": $info,
-  ),
-  "warning": (
-    "backgroundColor": scale-color($warning, $lightness: 80%),
-    "titleColor": #555,
-    "buttonColor": $warning,
-  ),
-  "error": (
-    "backgroundColor": scale-color($alert_error, $lightness: 80%),
-    "titleColor": #555,
-    "buttonColor": $alert_error,
-  ),
-  "text": (
-    "backgroundColor": $back_grey,
-    "titleColor": #000,
-    "buttonColor": $back_grey,
-  ),
-);
-
-.alert {
-  border-radius: 4px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 12px 16px;
-  gap: 8px;
-
-  @each $alertType in map-keys($alerts) {
-    &-#{$alertType} {
-      background-color: map-get(map-get($alerts, $alertType), backgroundColor);
-
-      .alertAction > button {
-        background-color: map-get(map-get($alerts, $alertType), buttonColor) !important;
-        border: 0 !important;
-        cursor: pointer;
-        transition: all 0.2s ease-in-out;
-        min-width: auto !important;
-      }
-
-      .alert-title {
-        color: map-get(map-get($alerts, $alertType), titleColor);
-        @include text-14-b;
-      }
-
-      .alert-close {
-        border-color: map-get(map-get($alerts, $alertType), backgroundColor);
-      }
-    }
-  }
-}
