--- src/components/Alert/MuiAlert.stories.tsx
@@ -1,42 +1,0 @@
-import InfoOutlinedIcon from '@mui/icons-material/InfoOutlined'
-import { StoryFn } from '@storybook/react'
-import React from 'react'
-
-import { IMuiAlertProps, MuiAlert as Alert } from './MuiAlert'
-
-export default {
-  title: 'MUI/Alert',
-  component: Alert,
-}
-
-const Template: StoryFn<IMuiAlertProps> = (args) => (
-  <Alert {...args} icon={<InfoOutlinedIcon />}>
-    You have 5 new offers and you have to click the button to able to accept them and also go to offers and download the
-    invoices
-  </Alert>
-)
-
-export const Green = Template.bind({})
-Green.args = {
-  variant: 'success',
-}
-
-export const Red = Template.bind({})
-Red.args = {
-  variant: 'error',
-}
-
-export const Blue = Template.bind({})
-Blue.args = {
-  variant: 'info',
-}
-
-export const Yellow = Template.bind({})
-Yellow.args = {
-  variant: 'warning',
-}
-
-export const Grey = Template.bind({})
-Grey.args = {
-  variant: 'grey',
-}
