--- src/components/Alert/index.ts
@@ -1,2 +1,0 @@
-export { default as Alert } from './Alert'
-export { MuiAlert } from './MuiAlert'
