+++ types/components/Alert/Alert.stories.d.ts
@@ -1,27 +1,8 @@
-import React from 'react';
-declare const _default: {
-    title: string;
-    component: React.FC<import("./Alert").AlertProps>;
-};
+import { AlertProps } from './Alert';
+declare const _default: import("@storybook/types").ComponentAnnotations<import("@storybook/react/dist/types-0a347bb9").R, import("@storybook/types").Args>;
 export default _default;
-export declare const Green: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, import("./Alert").AlertProps & {
-    children?: React.ReactNode;
-}>;
-export declare const Red: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, import("./Alert").AlertProps & {
-    children?: React.ReactNode;
-}>;
-export declare const Blue: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, import("./Alert").AlertProps & {
-    children?: React.ReactNode;
-}>;
-export declare const Yellow: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, import("./Alert").AlertProps & {
-    children?: React.ReactNode;
-}>;
-export declare const Grey: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, import("./Alert").AlertProps & {
-    children?: React.ReactNode;
-}>;
-export declare const WithButton: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, import("./Alert").AlertProps & {
-    children?: React.ReactNode;
-}>;
-export declare const Multiline: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, import("./Alert").AlertProps & {
-    children?: React.ReactNode;
-}>;
+export declare const Green: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, AlertProps>;
+export declare const Red: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, AlertProps>;
+export declare const Blue: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, AlertProps>;
+export declare const Yellow: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, AlertProps>;
+export declare const Grey: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, AlertProps>;
